+++ src/sampling.c
@@ -1,6 +1,6 @@
 /* Sampling with probability proportional to size (pps)
 
-   Copyright (C) 2024 Tobias Schoch (e-mail: tobias.schoch@gmail.com)
+   Copyright (C) 2024-2026 Tobias Schoch (e-mail: tobias.schoch@gmail.com)
 
    This library is free software; you can redistribute it and/or
    modify it under the terms of the GNU Library General Public